hide tweets cron if only 0 and moved columns to table file
